Move removeRemoteSource from RemoteTaskFactory to RemoteTask
