+++ presto-main/src/main/java/com/facebook/presto/execution/MemoryTrackingRemoteTaskFactory.java
@@ -22,7 +22,6 @@ import com.facebook.presto.metadata.Split;
 import com.facebook.presto.sql.planner.PlanFragment;
 import com.facebook.presto.sql.planner.plan.PlanNodeId;
 import com.google.common.collect.Multimap;
-import com.google.common.util.concurrent.ListenableFuture;
 
 import java.util.OptionalInt;
 
@@ -63,12 +62,6 @@ public class MemoryTrackingRemoteTaskFactory
 
         task.addStateChangeListener(new UpdatePeakMemory(stateMachine));
         return task;
-    }
-
-    @Override
-    public ListenableFuture<?> removeRemoteSource(TaskId taskId, TaskId remoteSourceTaskId)
-    {
-        return remoteTaskFactory.removeRemoteSource(taskId, remoteSourceTaskId);
     }
 
     private static final class UpdatePeakMemory
