+++ presto-main/src/main/java/com/facebook/presto/execution/RemoteTaskFactory.java
@@ -21,7 +21,6 @@ import com.facebook.presto.metadata.Split;
 import com.facebook.presto.sql.planner.PlanFragment;
 import com.facebook.presto.sql.planner.plan.PlanNodeId;
 import com.google.common.collect.Multimap;
-import com.google.common.util.concurrent.ListenableFuture;
 
 import java.util.OptionalInt;
 
@@ -36,6 +35,4 @@ public interface RemoteTaskFactory
             OutputBuffers outputBuffers,
             PartitionedSplitCountTracker partitionedSplitCountTracker,
             boolean summarizeTaskInfo);
-
-    ListenableFuture<?> removeRemoteSource(TaskId taskId, TaskId remoteSourceTaskId);
 }
